A small fix to make Windows C++ happy

+++ srtp/CryptoContext.cpp
@@ -395,7 +395,7 @@ bool CryptoContext::checkReplay(uint16_t newSeq)
 
         delta = -delta;
         int idx = (int)delta / 64;
-        uint64_t bit = 1UL << (delta % 64);
+        uint64_t bit = (uint64_t)1UL << (delta % 64);
         if ((replay_window[idx] & bit) == bit) {
             return false;  /* Packet already received ! */
         }
@@ -443,7 +443,7 @@ void CryptoContext::update(uint16_t newSeq)
     else {
         delta = -delta;
         int idx = (int)delta / 64;
-        uint64_t bit = 1UL << (delta % 64);
+        uint64_t bit = (uint64_t)1UL << (delta % 64);
         replay_window[idx] |= bit;
     }
